装饰模式实现
