+++ weixin-java-common/src/main/java/me/chanjar/weixin/common/util/http/MediaDownloadRequestExecutor.java
@@ -1,0 +1,161 @@
+package me.chanjar.weixin.common.util.http;
+
+import jodd.http.HttpConnectionProvider;
+import jodd.http.HttpRequest;
+import jodd.http.HttpResponse;
+import jodd.http.ProxyInfo;
+import me.chanjar.weixin.common.bean.result.WxError;
+import me.chanjar.weixin.common.exception.WxErrorException;
+import me.chanjar.weixin.common.util.fs.FileUtils;
+import me.chanjar.weixin.common.util.http.apache.InputStreamResponseHandler;
+import me.chanjar.weixin.common.util.http.apache.Utf8ResponseHandler;
+import org.apache.commons.lang3.StringUtils;
+import org.apache.http.Header;
+import org.apache.http.HttpHost;
+import org.apache.http.client.config.RequestConfig;
+import org.apache.http.client.methods.CloseableHttpResponse;
+import org.apache.http.client.methods.HttpGet;
+import org.apache.http.entity.ContentType;
+import org.apache.http.impl.client.CloseableHttpClient;
+
+import java.io.ByteArrayInputStream;
+import java.io.File;
+import java.io.IOException;
+import java.io.InputStream;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
+
+/**
+ * 下载媒体文件请求执行器，请求的参数是String, 返回的结果是File
+ * 视频文件不支持下载
+ *
+ * @author Daniel Qian
+ */
+public class MediaDownloadRequestExecutor implements RequestExecutor<File, String> {
+
+  private File tmpDirFile;
+
+  public MediaDownloadRequestExecutor(File tmpDirFile) {
+    this.tmpDirFile = tmpDirFile;
+  }
+
+  @Override
+  public File execute(RequestHttp requestHttp, String uri, String queryParam) throws WxErrorException, IOException {
+    if (requestHttp.getHttpClient() instanceof CloseableHttpClient) {
+      CloseableHttpClient httpClient = (CloseableHttpClient) requestHttp.getHttpClient();
+      HttpHost httpProxy = (HttpHost) requestHttp.getHttpProxy();
+      return executeApache(httpClient, httpProxy, uri, queryParam);
+    }
+    if (requestHttp.getHttpClient() instanceof HttpConnectionProvider) {
+      HttpConnectionProvider provider = (HttpConnectionProvider) requestHttp.getHttpClient();
+      ProxyInfo proxyInfo = (ProxyInfo) requestHttp.getHttpProxy();
+      return executeJodd(provider, proxyInfo, uri, queryParam);
+    } else {
+      //这里需要抛出异常，需要优化
+      return null;
+    }
+  }
+
+  private String getFileNameJodd(HttpResponse response) throws WxErrorException {
+    String content = response.header("Content-disposition");
+    if (content == null || content.length() == 0) {
+      throw new WxErrorException(WxError.newBuilder().setErrorMsg("无法获取到文件名").build());
+    }
+
+    Pattern p = Pattern.compile(".*filename=\"(.*)\"");
+    Matcher m = p.matcher(content);
+    if (m.matches()) {
+      return m.group(1);
+    }
+    throw new WxErrorException(WxError.newBuilder().setErrorMsg("无法获取到文件名").build());
+  }
+
+  private String getFileNameApache(CloseableHttpResponse response) throws WxErrorException {
+    Header[] contentDispositionHeader = response.getHeaders("Content-disposition");
+    if(contentDispositionHeader == null || contentDispositionHeader.length == 0){
+      throw new WxErrorException(WxError.newBuilder().setErrorMsg("无法获取到文件名").build());
+    }
+
+    Pattern p = Pattern.compile(".*filename=\"(.*)\"");
+    Matcher m = p.matcher(contentDispositionHeader[0].getValue());
+    if(m.matches()){
+      return m.group(1);
+    }
+    throw new WxErrorException(WxError.newBuilder().setErrorMsg("无法获取到文件名").build());
+  }
+
+
+  private File executeApache(CloseableHttpClient httpclient, HttpHost httpProxy, String uri, String queryParam) throws WxErrorException, IOException {
+    if (queryParam != null) {
+      if (uri.indexOf('?') == -1) {
+        uri += '?';
+      }
+      uri += uri.endsWith("?") ? queryParam : '&' + queryParam;
+    }
+
+    HttpGet httpGet = new HttpGet(uri);
+    if (httpProxy != null) {
+      RequestConfig config = RequestConfig.custom().setProxy(httpProxy).build();
+      httpGet.setConfig(config);
+    }
+
+    try (CloseableHttpResponse response = httpclient.execute(httpGet);
+         InputStream inputStream = InputStreamResponseHandler.INSTANCE
+           .handleResponse(response)) {
+
+      Header[] contentTypeHeader = response.getHeaders("Content-Type");
+      if (contentTypeHeader != null && contentTypeHeader.length > 0) {
+        if (contentTypeHeader[0].getValue().startsWith(ContentType.APPLICATION_JSON.getMimeType())) {
+          // application/json; encoding=utf-8 下载媒体文件出错
+          String responseContent = Utf8ResponseHandler.INSTANCE.handleResponse(response);
+          throw new WxErrorException(WxError.fromJson(responseContent));
+        }
+      }
+
+      String fileName = getFileNameApache(response);
+      if (StringUtils.isBlank(fileName)) {
+        return null;
+      }
+
+      String[] nameAndExt = fileName.split("\\.");
+      return FileUtils.createTmpFile(inputStream, nameAndExt[0], nameAndExt[1], this.tmpDirFile);
+
+    } finally {
+      httpGet.releaseConnection();
+    }
+
+  }
+
+
+  private File executeJodd(HttpConnectionProvider provider, ProxyInfo proxyInfo, String uri, String queryParam) throws WxErrorException, IOException {
+    if (queryParam != null) {
+      if (uri.indexOf('?') == -1) {
+        uri += '?';
+      }
+      uri += uri.endsWith("?") ? queryParam : '&' + queryParam;
+    }
+
+    HttpRequest request = HttpRequest.post(uri);
+    if (proxyInfo != null) {
+      provider.useProxy(proxyInfo);
+    }
+    request.withConnectionProvider(provider);
+    HttpResponse response = request.send();
+    String contentType = response.header("Content-Type");
+    if (contentType != null && contentType.startsWith("application/json")) {
+      // application/json; encoding=utf-8 下载媒体文件出错
+      throw new WxErrorException(WxError.fromJson(response.bodyText()));
+    }
+
+    String fileName = getFileNameJodd(response);
+    if (StringUtils.isBlank(fileName)) {
+      return null;
+    }
+
+    InputStream inputStream = new ByteArrayInputStream(response.bodyBytes());
+    String[] nameAndExt = fileName.split("\\.");
+    return FileUtils.createTmpFile(inputStream, nameAndExt[0], nameAndExt[1], this.tmpDirFile);
+  }
+
+
+}
