+++ weixin-java-common/src/main/java/me/chanjar/weixin/common/util/http/RequestExecutor.java
@@ -10,16 +10,14 @@ import java.io.IOException;
  * @param <T> 返回值类型
  * @param <E> 请求参数类型
  */
-public interface RequestExecutor<T, H, P, E> {
+public interface RequestExecutor<T, E> {
 
   /**
-   * @param httpClient
-   * @param httpProxy  http代理对象，如果没有配置代理则为空
-   * @param uri        uri
-   * @param data       数据
+   * @param uri  uri
+   * @param data 数据
    * @throws WxErrorException
    * @throws IOException
    */
-  T execute(H httpClient, P httpProxy, String uri, E data) throws WxErrorException, IOException;
+  T execute(RequestHttp requestHttp, String uri, E data) throws WxErrorException, IOException;
 
 }
