+++ weixin-java-common/src/main/java/me/chanjar/weixin/common/util/http/RequestHttp.java
@@ -1,0 +1,20 @@
+package me.chanjar.weixin.common.util.http;
+
+/**
+ * Created by ecoolper on 2017/4/22.
+ */
+public interface RequestHttp {
+
+  /**
+   * httpClient
+   * @return
+   */
+  Object getHttpClient();
+
+  /**
+   * httpProxy
+   * @return
+   */
+  Object getHttpProxy();
+
+}
