+++ weixin-java-common/src/main/java/me/chanjar/weixin/common/util/http/SimpleGetRequestExecutor.java
@@ -1,0 +1,91 @@
+package me.chanjar.weixin.common.util.http;
+
+import jodd.http.HttpConnectionProvider;
+import jodd.http.HttpRequest;
+import jodd.http.HttpResponse;
+import jodd.http.ProxyInfo;
+import me.chanjar.weixin.common.bean.result.WxError;
+import me.chanjar.weixin.common.exception.WxErrorException;
+import me.chanjar.weixin.common.util.http.apache.Utf8ResponseHandler;
+import org.apache.http.HttpHost;
+import org.apache.http.client.config.RequestConfig;
+import org.apache.http.client.methods.CloseableHttpResponse;
+import org.apache.http.client.methods.HttpGet;
+import org.apache.http.impl.client.CloseableHttpClient;
+
+import java.io.IOException;
+
+/**
+ * 简单的GET请求执行器，请求的参数是String, 返回的结果也是String
+ *
+ * @author Daniel Qian
+ */
+public class SimpleGetRequestExecutor implements RequestExecutor<String, String> {
+
+  @Override
+  public String execute(RequestHttp requestHttp, String uri, String queryParam) throws WxErrorException, IOException {
+    if (requestHttp.getHttpClient() instanceof CloseableHttpClient) {
+      CloseableHttpClient httpClient = (CloseableHttpClient) requestHttp.getHttpClient();
+      HttpHost httpProxy = (HttpHost) requestHttp.getHttpProxy();
+      return executeApache(httpClient, httpProxy, uri, queryParam);
+    }
+    if (requestHttp.getHttpClient() instanceof HttpConnectionProvider) {
+      HttpConnectionProvider provider = (HttpConnectionProvider) requestHttp.getHttpClient();
+      ProxyInfo proxyInfo = (ProxyInfo) requestHttp.getHttpProxy();
+      return executeJodd(provider, proxyInfo, uri, queryParam);
+    } else {
+      //这里需要抛出异常，需要优化
+      return null;
+    }
+  }
+
+
+  private String executeApache(CloseableHttpClient httpclient, HttpHost httpProxy, String uri, String queryParam) throws WxErrorException, IOException {
+    if (queryParam != null) {
+      if (uri.indexOf('?') == -1) {
+        uri += '?';
+      }
+      uri += uri.endsWith("?") ? queryParam : '&' + queryParam;
+    }
+    HttpGet httpGet = new HttpGet(uri);
+    if (httpProxy != null) {
+      RequestConfig config = RequestConfig.custom().setProxy(httpProxy).build();
+      httpGet.setConfig(config);
+    }
+
+    try (CloseableHttpResponse response = httpclient.execute(httpGet)) {
+      String responseContent = Utf8ResponseHandler.INSTANCE.handleResponse(response);
+      WxError error = WxError.fromJson(responseContent);
+      if (error.getErrorCode() != 0) {
+        throw new WxErrorException(error);
+      }
+      return responseContent;
+    } finally {
+      httpGet.releaseConnection();
+    }
+  }
+
+
+  private String executeJodd(HttpConnectionProvider provider, ProxyInfo proxyInfo, String uri, String queryParam) throws WxErrorException, IOException {
+    if (queryParam != null) {
+      if (uri.indexOf('?') == -1) {
+        uri += '?';
+      }
+      uri += uri.endsWith("?") ? queryParam : '&' + queryParam;
+    }
+
+    HttpRequest request = HttpRequest.get(uri);
+    if (proxyInfo != null) {
+      provider.useProxy(proxyInfo);
+    }
+    request.withConnectionProvider(provider);
+    HttpResponse response = request.send();
+    String responseContent = response.bodyText();
+    WxError error = WxError.fromJson(responseContent);
+    if (error.getErrorCode() != 0) {
+      throw new WxErrorException(error);
+    }
+    return responseContent;
+  }
+
+}
