--- weixin-java-common/src/main/java/me/chanjar/weixin/common/util/http/jodd/SimpleGetRequestExecutor.java
@@ -1,43 +1,0 @@
-package me.chanjar.weixin.common.util.http.jodd;
-
-import jodd.http.HttpConnectionProvider;
-import jodd.http.HttpRequest;
-import jodd.http.HttpResponse;
-import jodd.http.ProxyInfo;
-import me.chanjar.weixin.common.bean.result.WxError;
-import me.chanjar.weixin.common.exception.WxErrorException;
-import me.chanjar.weixin.common.util.http.RequestExecutor;
-
-import java.io.IOException;
-
-/**
- * 简单的GET请求执行器，请求的参数是String, 返回的结果也是String
- *
- * @author Daniel Qian
- */
-public class SimpleGetRequestExecutor implements RequestExecutor<String, HttpConnectionProvider, ProxyInfo, String> {
-
-  @Override
-  public String execute(HttpConnectionProvider provider, ProxyInfo httpProxy, String uri, String queryParam) throws WxErrorException, IOException {
-    if (queryParam != null) {
-      if (uri.indexOf('?') == -1) {
-        uri += '?';
-      }
-      uri += uri.endsWith("?") ? queryParam : '&' + queryParam;
-    }
-
-    HttpRequest request = HttpRequest.get(uri);
-    if (httpProxy != null) {
-      provider.useProxy(httpProxy);
-    }
-    request.withConnectionProvider(provider);
-    HttpResponse response = request.send();
-    String responseContent = response.bodyText();
-    WxError error = WxError.fromJson(responseContent);
-    if (error.getErrorCode() != 0) {
-      throw new WxErrorException(error);
-    }
-    return responseContent;
-  }
-
-}
